Wiki placeholders (#29)
Depends on #28.

This PR adds the infrastructure for various kinds of placeholders.
Firstly, inline code wasn't handled correctly before, and `<br>` in
inline code would produce a line break.

The goal is for us to write `#rules` and `:stackoverflow:` in the
articles, and for this to turn into the correct emotes and channel
mentions.