Refactor anti-screenshot (#18)
This basically just pulls out the creation of the embed and removes
`create_embed`, which was only used in one place, and didn't have much
of a right to exist.

I initially wanted to fix #12, but was unable to locate the source of
the problem. Perhaps the screenshot component is duplicated in
`wheatley-private`?