Add users current score to the scoreboard (#87)
Currently there is no manner with The Button to be able to tell what
your current score is without pressing the button. This minor addition
adds your current score to the scoreboard button when pressed.

I've kept this change pretty minor, but another possibility is to also
display your current ranking. I've though for the time being kept the
change small.