Refactor logging (#103)
These changes move the `critical_error()`, `ignorable_error()`, and
`milestone()` functions into the `Wheatley` object and use a log thread
channel to output alerts instead of sending DMs to Zelis. This also
eliminates the global `client` and `zelis` hackery.

Some places had hardcoded DMs to Zelis; this was replaced with a generic
`alert()` function that also logs to the log thread.

The goal here is to take a first step towards decoupling Wheatley at
least a little bit from TCCPP-specifics, and make it at least a tiny bit
easier to run independent instances of the bot.

One notable change is that failures in `main.ts` are now only logged to
the normal stdout log since these happen prior to a `Wheatley` object
having been created. It is arguably not really necessary to get a
discord notification whenever the bot failed to start; whoever started
the bot should notice right away anyways, so this should be alright?