Implemented feature to not fill up all reaction spots in #memes (#131)
I created a feature so that not all of the reaction spots cannot be
filled up unless one of them is a negative reaction. This is to prevent
users from filling up all spots from a delet_this or another negative
reaction.

Created one new function `should_delete_reaction()` to check if it
should be deleted

I also fixed a typo in the Makefile.