Refactor channel access (#169)
Instead of having a massive map of channels in wheatley, which was
convenient but not very conducive to running the bot elsewhere, it's
more principled to have every module just fetch their own channels as
part of startup. This PR implements that. These changes were also mostly
done by claude.

This also fixes the horrible typescript errors that manifested in
58d6ef708e74f7d6f503206d33fbea5fec422ed8 where a mundane change in an
unrelated part of the codebase resulted in errors about union sizes,
caused because of the massive channel type in wheatley.

Twin PR: https://github.com/TCCPP/wheatley-private/pull/15